fix: block align inodes

+++ solution/fs_utils.h
@@ -4,7 +4,7 @@
 #include "wfs.h"
 #include <stddef.h>
 
-int calculate_required_size(size_t inode_count, size_t data_block_count);
+size_t calculate_required_size(size_t inode_count, size_t data_block_count);
 
 int initialize_disk(const char *disk_file, size_t inode_count,
                     size_t data_block_count, int required_size);
